refactor: move eval into a separate `primer-eval` package
