+++ primer-eval/src/Primer/Eval/ProgError.hs
@@ -5,6 +5,7 @@ import Foreword
 import Data.Aeson (FromJSON (..), ToJSON (..))
 import Primer.Action.Errors (ActionError)
 import Primer.Core.Meta (GVarName, ID, ModuleName, TyConName, TyVarName, ValConName)
+import Primer.Eval.EvalError (EvalError)
 import Primer.JSON (CustomJSON (..), PrimerJSON)
 import Primer.Name (Name)
 
@@ -30,6 +31,7 @@ data ProgError
   | TyConParamClash Name
   | ValConParamClash Name
   | ActionError ActionError
+  | EvalError EvalError
   | -- | Currently copy/paste is only exposed in the frontend via select
     --   channels, which should never go wrong. Consequently, this is an
     --   "internal error" which should never happen!
